factor: Move the Miller-Rabin primality test to its own module.
